--- a/src/core/ctxt.ml
+++ b/src/core/ctxt.ml
@@ -33,6 +33,20 @@ let to_prod : ctxt -> term -> term * int = fun ctx t ->
   let t, c = List.fold_left f (lift t, 0) ctx in
   Bindlib.unbox t, c
 
+(** [to_prod_box bctx t] is similar to [to_prod bctx t] but operates on boxed
+    contexts and terms. *)
+let to_prod_box : bctxt -> tbox -> tbox * int = fun bctx t ->
+  let f (t, c) (x, a) =
+    let b = Bindlib.bind_var x t in
+    (_Prod a b, c + 1)
+  in
+  List.fold_left f (t, 0) bctx
+
+(** [box_context ctx] lifts context [ctx] to a boxed context. *)
+let box_context : ctxt -> bctxt =
+  List.filter_map
+    (fun (x, t, u) -> if u = None then Some (x, lift t) else None)
+
 (** [to_abst ctx t] builds a sequence of abstractions over the context [ctx],
     in the term [t]. *)
 let to_abst : ctxt -> term -> term = fun ctx t ->