--- a/src/core/infer.ml
+++ b/src/core/infer.ml
@@ -1,234 +1,189 @@
 (** Type inference and checking *)
 
-open Timed
 open Common
-open Error
-open Term
-open Print
-open Debug
 open Lplib
+open Term
+open Timed
 
 (** Logging function for typing. *)
-let log_infr = Logger.make 'i' "infr" "type inference/checking"
-let log_infr = log_infr.pp
-
-(** Given a meta [m] of type [Πx1:a1,..,Πxn:an,b], [set_to_prod d p m] sets
-   [m] to a product term of the form [Πy:m1[x1;..;xn],m2[x1;..;xn;y]] with
-   [m1] and [m2] fresh metavariables, and adds these metavariables to [p].
-   [d] is the call depth used for debug. *)
-let set_to_prod : int -> problem -> meta -> unit = fun d p m ->
-  let n = m.meta_arity in
-  let env, s = Env.of_prod_nth [] n !(m.meta_type) in
-  let vs = Env.vars env in
-  let xs = Array.map _Vari vs in
-  (* domain *)
-  let u1 = Env.to_prod env _Type in
-  let m1 = LibMeta.fresh p u1 n in
-  let a = _Meta m1 xs in
-  (* codomain *)
-  let y = new_tvar "y" in
-  let env' = Env.add y (_Meta m1 xs) None env in
-  let u2 = Env.to_prod env' (lift s) in
-  let m2 = LibMeta.fresh p u2 (n+1) in
-  let b = Bindlib.bind_var y (_Meta m2 (Array.append xs [|_Vari y|])) in
-  (* result *)
-  let r = _Prod a b in
-  if Logger.log_enabled () then
-    log_infr (Color.red "%a%a ≔ %a") D.depth d
-      pp_meta m pp_term (Bindlib.unbox r);
-  LibMeta.set p m (Bindlib.unbox (Bindlib.bind_mvar vs r))
-
-(** [conv d p c a b] adds the the constraint [(c,a,b)] in [p], if [a] and
-   [b] are not convertible. [d] is the call depth used for debug. *)
-let conv : int -> problem -> ctxt -> term -> term -> unit = fun d p c a b ->
-  if not (Eval.pure_eq_modulo c a b) then begin
-    let cstr = (c,a,b) in
-    if Logger.log_enabled () then
-      log_infr (Color.mag "%aadd %a") D.depth d pp_constr cstr;
-    p := {!p with to_solve = cstr::!p.to_solve}
-  end
+let log = Logger.make 'i' "infr" "type inference/checking"
+let log = log.pp
 
 (** Exception that may be raised by type inference. *)
 exception NotTypable
 
-(** [infer d p c t] tries to infer a type for the term [t] in context [c],
-   possibly adding new unification constraints in [p]. The set of
-   metavariables of [p] are updated if some metavariables are instantiated or
-   created. The returned type is well-sorted if the recorded constraints are
-   satisfied. [c] must be well sorted. [d] is the call depth used for debug.
-@raise NotTypable when the term is not typable (when encountering an
-   abstraction over a kind). *)
-let rec infer : int -> problem -> ctxt -> term -> term = fun d p c t ->
-  let return v =
-    if Logger.log_enabled () then log_infr "%agot %a" D.depth d pp_term v; v
+(** [unif pb c a b] solves the unification problem [c ⊢ a ≡ b]. Current
+    implementation collects constraints in {!val:constraints} then solves
+    them at the end of type checking. *)
+let unif : problem -> ctxt -> term -> term -> unit =
+ fun pb c a b ->
+ if not (Eval.pure_eq_modulo c a b) then
+ (* NOTE: eq_modulo is used because the unification algorithm counts on
+    the fact that no constraint is added in some cases (see test
+    "245b.lp"). We may however want to reduce the number of calls to
+    [eq_modulo]. *)
+   begin
+     if Logger.log_enabled () then
+       log (Color.yel "add constraint %a") Print.pp_constr
+         (c, a, b);
+     pb := {!pb with to_solve = (c, a, b) :: !pb.to_solve}
+   end
+
+(** {1 Handling coercions} *)
+
+(* Simply unify types, no coercions yet. *)
+let coerce pb c t a b = unif pb c a b; t
+
+(** {1 Other rules} *)
+
+(** [type_enforce pb c a] returns a tuple [(a',s)] where [a'] is refined
+    term [a] and [s] is a sort (Type or Kind) such that [a'] is of type
+    [s]. *)
+let rec type_enforce : problem -> ctxt -> term -> term * term =
+ fun pb c a ->
+  if Logger.log_enabled () then log "Type enforce [%a]" Print.pp_term a;
+  let a, s = infer pb c a in
+  let sort =
+    match unfold s with
+    | Kind -> mk_Kind
+    | Type -> mk_Type
+    | _ -> mk_Type
+    (* FIXME The algorithm should be able to backtrack on the choice of
+       this sort, first trying [Type], and if it does not succeed, trying
+       [Kind]. *)
   in
-  if Logger.log_enabled () then
-    log_infr "%ainfer %a%a" D.depth d pp_ctxt c pp_term t;
+  let a = coerce pb c a s sort in
+  (a, sort)
 
+(** [force pb c t a] returns a term [t'] such that [t'] has type [a],
+    and [t'] is the refinement of [t]. *)
+and force : problem -> ctxt -> term -> term -> term =
+ fun pb c te ty ->
+ if Logger.log_enabled () then
+   log "Force [%a] of [%a]" Print.pp_term te Print.pp_term ty;
+ let (t, a) = infer pb c te in
+ coerce pb c t a ty
+
+and infer_aux : problem -> ctxt -> term -> term * term =
+ fun pb c t ->
   match unfold t with
-  | Patt(_,_,_) -> assert false (* Forbidden case. *)
-  | TEnv(_,_)   -> assert false (* Forbidden case. *)
-  | Kind        -> assert false (* Forbidden case. *)
-  | Wild        -> assert false (* Forbidden case. *)
-  | TRef(_)     -> assert false (* Forbidden case. *)
-
-  (* -------------------
-      c ⊢ Type ⇒ Kind  *)
-  | Type -> return mk_Kind
-
-  (* ---------------------------------
-      c ⊢ Vari x ⇒ Ctxt.type_of x c  *)
-  | Vari x -> (try return (Ctxt.type_of x c) with Not_found -> assert false)
-
-  (* -------------------------------
-      c ⊢ Symb s ⇒ !(s.sym_type)  *)
-  | Symb s -> return !(s.sym_type)
-
-  (*  c ⊢ a ⇐ Type    c, x:a ⊢ b ⇒ s in {Type,Kind}
-     -----------------------------------------
-                c ⊢ Prod(a,b) ⇒ s            *)
-  | Prod(a,b) ->
-      check (d+1) p c a mk_Type;
-      let _,b,c' = Ctxt.unbind c a None b in
-      let s = infer (d+1) p c' b in
-      begin match unfold s with
-      | (Type | Kind) as s -> s
-      | _ -> conv d p c' s mk_Type; mk_Type
-      (* Here, we force [s] to be equivalent to [Type] as there is little
-         (no?) chance that it can be a kind. *)
-      end
-
-  (*  c ⊢ a ⇐ Type    c, x:a ⊢ t ⇒ b   b ≠ Kind
-     --------------------------------------------
-             c ⊢ Abst(a,t) ⇒ Prod(a,b)          *)
-  | Abst(a,t) ->
-      check (d+1) p c a mk_Type;
-      let x,t,c' = Ctxt.unbind c a None t in
-      let b = infer (d+1) p c' t in
-      begin match unfold b with
-      | Kind ->
-          wrn None "Abstraction on [%a] is not allowed." Print.pp_term t;
-          raise NotTypable
-      | _ -> return (mk_Prod(a, Bindlib.unbox (Bindlib.bind_var x (lift b))))
-      end
-
-  (*  c ⊢ t ⇒ Prod(a,b)    c ⊢ u ⇐ a
-     ------------------------------------
-         c ⊢ Appl(t,u) ⇒ subst b u      *)
-  | Appl(t,u) ->
-      (* [get_prod f typ] returns the domain and codomain of [t] if [t] is a
-         product. If [t] is a metavariable, then it instantiates it with a
-         product and calls [get_prod f typ] again. Otherwise, it calls [f
-         typ]. *)
-      let rec get_prod f typ =
-        match unfold typ with
-        | Prod(a,b) -> (a,b)
-        | Meta(m,_) -> set_to_prod d p m; get_prod f typ
-        | _ -> f typ
+  | Patt _ -> assert false
+  | TEnv _ -> assert false
+  | Kind -> assert false
+  | Wild -> assert false
+  | TRef _ -> assert false
+  | Type -> (mk_Type, mk_Kind)
+  | Vari x ->
+      let a = try Ctxt.type_of x c with Not_found -> assert false in
+      (t, a)
+  | Symb s -> (t, !(s.sym_type))
+  (* All metavariables inserted are typed. *)
+  | (Meta (m, ts)) as t ->
+      let rec ref_esubst i range =
+        (* Refine terms of the explicit substitution. *)
+        if i >= Array.length ts then range else
+          match unfold range with
+          | Prod(ai, b) ->
+              ts.(i) <- force pb c ts.(i) ai;
+              let b = Bindlib.subst b ts.(i) in
+              ref_esubst (i + 1) b
+          | _ ->
+              (* Meta type must be a product of arity greater or equal
+                 to the environment *)
+              assert false
       in
-      let get_prod_whnf = (* assumes that its argument is in whnf *)
-        get_prod (fun typ ->
-            let a = LibMeta.make p c mk_Type in
-            (* We force [b] to be of type [Type] as there is little (no?)
-               chance that it can be a kind. *)
-            let b = LibMeta.make_codomain p c a in
-            conv d p c typ (mk_Prod(a,b)); (a,b)) in
-      let get_prod =
-        get_prod (fun typ -> get_prod_whnf (Eval.whnf c typ)) in
-      let a, b = get_prod (infer (d+1) p c t) in
-      check (d+1) p c u a;
-      return (Bindlib.subst b u)
+      let range = ref_esubst 0 !(m.meta_type) in
+      (t, range)
+  | LLet (t_ty, t, u) ->
+      (* Check that [a] is a type, and refine it. *)
+      let t_ty, _ = type_enforce pb c t_ty in
+      (* Check that [t] is of type [t_ty], and refine it *)
+      let t = force pb c t t_ty in
+      (* Unbind [u] and get new context with [x: t_ty ≔ t] *)
+      let x, u, c = Ctxt.unbind c t_ty (Some t) u in
+      (* Infer type of [u'] and refine it. *)
+      let u, u_ty = infer pb c u in
+      ( match unfold u_ty with
+        | Kind ->
+            Error.fatal_msg "Let bindings cannot have a body of type Kind.";
+            Error.fatal_msg "Body of let binding [%a] has type Kind."
+              Print.pp_term u;
+            raise NotTypable
+        | _ -> () );
+      let u_ty = Bindlib.(u_ty |> lift |> bind_var x |> unbox) in
+      let u = Bindlib.(u |> lift |> bind_var x |> unbox) in
+      (mk_LLet (t_ty, t, u), mk_LLet(t_ty, t, u_ty))
+  | Abst (dom, b) ->
+      (* Domain must by of type Type (and not Kind) *)
+      let dom = force pb c dom mk_Type in
+      let x, b, c = Ctxt.unbind c dom None b in
+      let b, range = infer pb c b in
+      let b = Bindlib.(lift b |> bind_var x |> unbox) in
+      let range = Bindlib.(lift range |> bind_var x |> unbox) in
+      (mk_Abst (dom, b), mk_Prod (dom, range))
+  | Prod (dom, b) ->
+      (* Domain must by of type Type (and not Kind) *)
+      let dom = force pb c dom mk_Type in
+      let x, b, c = Ctxt.unbind c dom None b in
+      let b, b_s = type_enforce pb c b in
+      let b = Bindlib.(lift b |> bind_var x |> unbox) in
+      (mk_Prod (dom, b), b_s)
+  | Appl (t, u) -> (
+      let t, t_ty = infer pb c t in
+      match Eval.whnf c t_ty with
+      | Prod (dom, b) ->
+          if Logger.log_enabled () then log "Appl-prod arg [%a]" Print.pp_term u;
+          let u = force pb c u dom in
+          (mk_Appl (t, u), Bindlib.subst b u)
+      | Meta (_, _) ->
+          let u, u_ty = infer pb c u in
+          let range = LibMeta.make_codomain pb c u_ty in
+          unif pb c t_ty (mk_Prod (u_ty, range));
+          (mk_Appl (t, u), Bindlib.subst range u)
+      | t_ty ->
+          let domain = LibMeta.make pb c mk_Type in
+          let range = LibMeta.make_codomain pb c domain in
+          let t = coerce pb c t t_ty (mk_Prod (domain, range)) in
+          if Logger.log_enabled () then log "Appl-default arg [%a]" Print.pp_term u;
+          let u = force pb c u domain in
+          (mk_Appl (t, u), Bindlib.subst range u) )
 
-  (* c ⊢ t ⇐ a   c, x:a ≔ t ⊢ u ⇒ b   c, x:a ≔ t ⊢ b : s in {Type,Kind}
-     ------------------------------------------------------------------
-        c ⊢ let x:a ≔ t in u ⇒ let x:a ≔ t in b
-
-     See Pure type systems with definitions, P. Severi and E. Poll,
-     LFCS 1994, http://doi.org/10.1007/3-540-58140-5_30. *)
-  | LLet(a,t,u) ->
-      check (d+1) p c a mk_Type;
-      check (d+1) p c t a;
-      let x,u,c' = Ctxt.unbind c a (Some t) u in
-      let b = infer (d+1) p c' u in
-      begin match unfold b with
-      | Kind ->
-          wrn None "Abstraction on [%a] is not allowed." Print.pp_term u;
-          raise NotTypable
-      | _ ->
-          let s = infer (d+1) p c' b in
-          begin match unfold s with
-          | Type | Kind -> ()
-          | _ -> conv d p c' s mk_Type
-          (* Here, we force [s] to be equivalent to [Type] as there is little
-             (no?) chance that it can be a kind. *)
-          end;
-          let b = Bindlib.unbox (Bindlib.bind_var x (lift b)) in
-          return (mk_LLet(a,t,b))
-      end
-
-  (*  c ⊢ term_of_meta m ts ⇒ a
-     ----------------------------
-         c ⊢ Meta(m,ts) ⇒ a      *)
-  | Meta(m,ts) ->
-      (* The type of [Meta(m,ts)] is the same as the one obtained by applying
-         to [ts] a new symbol having the same type as [m]. *)
-      let s = Term.create_sym (Sign.current_path()) Privat Const
-          Eager true (Pos.none ("?" ^ LibMeta.name m)) !(m.meta_type) [] in
-      if Logger.log_enabled () then
-        log_infr "%areplace meta by fresh symbol" D.depth d;
-      infer d p c
-        (Array.fold_left (fun acc t -> mk_Appl(acc,t)) (mk_Symb s) ts)
-
-(** [check d c t a] checks that the term [t] has type [a] in context [c],
-   possibly adding new unification constraints in [p]. [c] must be
-   well-formed and [a] well-sorted. [d] is the call depth used for debug.
-@raise NotTypable when the term is not typable (when encountering an
-   abstraction over a kind). *)
-and check : int -> problem -> ctxt -> term -> term -> unit = fun d p c t a ->
+and infer : problem -> ctxt -> term -> term * term = fun pb c t ->
+  if Logger.log_enabled () then log "Infer [%a]" Print.pp_term t;
+  let t, t_ty = infer_aux pb c t in
   if Logger.log_enabled () then
-    log_infr "%acheck %a" D.depth d pp_typing (c, t, a);
-  conv d p c (infer (d+1) p c t) a
+    log "Inferred [%a: %a]" Print.pp_term t Print.pp_term t_ty;
+  (t, t_ty)
 
-let infer =
-  let open Stdlib in let r = ref mk_Kind in fun d p c t ->
-  Debug.(record_time Typing (fun () -> r := infer d p c t)); !r
 
-let check d p c t a = Debug.(record_time Typing (fun () -> check d p c t a))
-
-(** [infer_noexn p c t] returns [None] if the type of [t] in context [c]
-   cannot be inferred, or [Some a] where [a] is some type of [t] in the
-   context [c], possibly adding new constraints in [p]. The metavariables of
-   [p] are updated when a metavariable is instantiated or created. [c] must
-   be well sorted. *)
-let infer_noexn : problem -> ctxt -> term -> term option = fun p c t ->
+(** [noexn f cs c args] initialises {!val:constraints} to [cs],
+    calls [f c args] and returns [Some(r,cs)] where [r] is the value of
+    the call to [f] and [cs] is the list of constraints gathered by
+    [f]. Function [f] may raise [NotTypable], in which case [None] is
+    returned. *)
+let noexn : (problem -> ctxt -> 'a -> 'b) -> problem -> ctxt -> 'a ->
+  'b option =
+  fun f pb c args ->
   try
-    if Logger.log_enabled () then
-      log_hndl (Color.blu "infer_noexn %a%a") pp_ctxt c pp_term t;
-    let a = time_of "infer" (fun () -> infer 0 p c t) in
-    if Logger.log_enabled () then
-      log_hndl (Color.blu "@[<v2>@[result of infer_noexn:@ %a@]%a@]")
-        pp_term a
-        (fun fmt constraints -> if constraints <> [] then
-          Format.fprintf fmt ";@ @[constraints:@ %a@]" pp_constrs constraints)
-        !p.to_solve;
-    Some a
+    Some (f pb c args)
   with NotTypable -> None
 
-(** [check_noexn p c t a] tells whether the term [t] has type [a] in the
-   context [c], possibly adding new constraints in |p]. The metavariables of
-   [p] are updated when a metavariable is instantiated or created. The context
-   [c] and the type [a] must be well sorted. *)
-let check_noexn : problem -> ctxt -> term -> term -> bool = fun p c t a ->
-  try
-    if Logger.log_enabled () then
-      log_hndl (Color.blu "check_noexn %a") pp_typing (c, t, a);
-    time_of "check" (fun () -> check 0 p c t a);
-    if Logger.log_enabled () && !p.to_solve <> [] then
-      log_hndl (Color.blu "result of check_noexn:%a") pp_constrs !p.to_solve;
-    true
-  with NotTypable -> false
+let infer_noexn pb c t =
+  if Logger.log_enabled () then
+    log "Top infer %a%a" Print.pp_ctxt c Print.pp_term t;
+  noexn infer pb c t
 
-(** Given a meta [m] of type [Πx1:a1,..,Πxn:an,b], [set_to_prod p m] sets
-   [m] to a product term of the form [Πy:m1[x1;..;xn],m2[x1;..;xn;y]] with
-   [m1] and [m2] fresh metavariables, and adds these metavariables to [p]. *)
-let set_to_prod = set_to_prod 0
+let check_noexn pb c t a =
+  if Logger.log_enabled () then log "Top check \"%a\"" Print.pp_typing
+      (c, t, a);
+  let force pb c (t, a) = force pb c t a in
+  noexn force pb c (t, a)
+
+let check_sort_noexn pb c t : (term * term) option =
+  if Logger.log_enabled () then
+    log "Top check sort %a%a" Print.pp_ctxt c Print.pp_term t;
+  noexn type_enforce pb c t
+
+let infer_noexn pb c t = Option.map snd (infer_noexn pb c t)
+let check_noexn pb c t a = check_noexn pb c t a <> None
+let check_sort_noexn pb c t = check_sort_noexn pb c t <> None