--- a/src/core/infer.mli
+++ b/src/core/infer.mli
@@ -2,11 +2,6 @@
 
 open Term
 
-(** Given a meta [m] of type [Πx1:a1,..,Πxn:an,b], [set_to_prod p m] sets [m]
-   to a product term of the form [Πy:m1[x1;..;xn],m2[x1;..;xn;y]] with fresh
-   metavariables [m1] and [m2], and updates the metas of the problem [p]. *)
-val set_to_prod : problem -> meta -> unit
-
 (** [infer_noexn p ctx t] returns [None] if the type of [t] in context [ctx]
    cannot be inferred, or [Some a] where [a] is some type of [t] in the
    context [ctx], possibly adding new constraints in [p]. The metavariables of
@@ -19,3 +14,5 @@ val infer_noexn : problem -> ctxt -> term -> term option
    [p] are updated when a metavariable is instantiated or created. The context
    [ctx] and the type [a] must be well sorted. *)
 val check_noexn : problem -> ctxt -> term -> term -> bool
+
+val check_sort_noexn : problem -> ctxt -> term -> bool