… CHANGES.md
… doc/terms.rst
… src/core/libMeta.ml
… src/core/sign.ml
… src/core/term.ml
… src/core/term.mli
… src/core/unif.ml
… src/export/dk.ml
… src/export/hrs.ml
… src/export/xtc.ml
… src/handle/command.ml
… src/handle/proof.ml
… src/handle/query.ml
… src/handle/query.mli
… src/handle/rewrite.ml
… src/handle/tactic.ml
… src/parsing/lpLexer.ml
… src/parsing/lpParser.mly
… src/parsing/pretty.ml
… src/parsing/scope.ml
… src/parsing/scope.mli
… src/parsing/syntax.ml
… src/tool/sr.ml
… tests/OK/767.lp
… tests/OK/admit.lp
… tests/OK/natproofs.lp
… tests/OK/perf_rw_fib-nonRightLin.lp
… tests/OK/rewrite1.lp
… tests/OK/tautologies.lp
… tests/OK/triangular.lp