This PR is equivalent to #563 but commits are separated in a more intelligible way:

1. Implementation of the basis of the new typechecker
2. Modifying signatures to make the type checker bidirectional (allowed to changed the type checked term), and subsequently modify calls to the typechecker
3. Introduce placeholders and generate meta-variables during type checking: placeholders of the form "?" are replaced by meta-variables during typechecking (or fully instantiated terms).

## Changelog

Introduce a new type checker which is able to refine (aka modify) terms. The algorithm is taken from <https://arxiv.org/abs/1202.4905>.

### Added

- Tems may now be _placeholders_. Placeholders are holes in the concrete syntax. They are refined into metavariables. *Placeholders cannot appear nonlinearly in terms*. From [A Bidirectional Refinement Algorithm...](https://arxiv.org/abs/1202.4905), p. 31,
  > Non linear placeholders are not allowed since two occurrences could be in contexts that bind different set of variables and instantiation with terms that live in one context would make no sense in the other one.

### Removed

- Unused variable warning: whether a variable is used or not cannot be decided while scoping anymore since placeholders that do not depend on variables may be refined later into metavariables that may depend on them.
- Placeholders do not depend on variables, hence the syntax `?M.[x;y]` is obsolete, only `?M` is valid

### Changed

- Naming metavariable is now useless since they must be linear.
- Because placeholders are simple holes, the term `_ → _` is scoped into a full dependent product `Π x: M, N` where `N` is a metavariable that depend on `x` (see file `tests/OK/767.lp`)
- Type checking takes longer because of refinement (not only the type must be destructured and rebuilt),
  |         | master | refiner |
  |---------|--------|---------|
  | holide  | 7:0    | 11:33   |
  | iprover | 5:58   | 6:50    |
- scoping term is more straightforward (no metavariable fetching), thus `scope_term` simplified