@gabrielhdt 2021-06-01
Although the coercions are not yet implemented, the code can be integrated. It provides a new typechecking algorithm, and it adds placeholders so that meta variables are generated at type check time.