@fblanqui 2021-06-01
Hello @gabrielhdt . There are a number of problems with this PR:
1) https://github.com/fblanqui/lib does not compile with it.
Remark: it's a good habit to check this library in addition to make tests. Perhaps we should integrate it in the CI.
2) Performances are degraded:
comparison between master and your PR on my machine:
make tests: 7.7s -> 9.2s = +20%
consider tests/OK/649_fo_27.lp and increase the number:
fo_45.lp: 0.643s -> 1.909s = +200%