@fblanqui 2021-06-01
@gabrielhdt It would be nice that you explained in more details what are the differences between the type inference/checking algorithm in master and the one in your PR. At first glance, I do not see much differences.
In fact, in your PR, I see two things:
1) At scoping time, replace underscores and implicit arguments by place holders. During type inference/checking, replace place holders by metavariables. You do this so that metavariables get well-typed contexts, as it may be necessary to add coercions in contexts too.
2) Perhaps some modifications in the type inference/checking algorithm. But this is not clear to me yet.
So, perhaps, this PR could be split in two. This would help understand the lost in performance.