@gabrielhdt 2021-06-01
Hi,
>     1. https://github.com/fblanqui/lib does not compile with it.
>        Remark: it's a good habit to check this library in addition to make tests. Perhaps we should integrate it in the CI.

This can be fixed by adding a final `reflexivity` to the proof script, line 224 of `Nat.lp`. I'll check whether this is justified or if it's a bug.