@fblanqui 2021-06-01
[fo_45.txt](https://github.com/Deducteam/lambdapi/files/6578464/fo_45.txt)