@gabrielhdt 2021-06-02
With first commit (so only new typechecker), `fo_45.lp` takes 1.1s (user time)