@gabrielhdt 2021-06-06
Regarding performance problem, I think they  are due to the fact that during typechecking, we traverse the context more often than before (twice as many times), since it is traversed to transform placeholders into meta variables and then to instantiate metavariables.

Things should be better if we change the type of meta variables, namely reduce the `meta` type to
```ocaml
type meta = { meta_key : int; meta_name : string option; meta_value : tmbinder ref }
```
and extend the type of problems to
```ocaml
type problem_aux =
  { ...
  ; metas : (ctxt * term) MetaMap.t }
```
so that the `arity` and `meta_type` field are moved to the context and type of the meta map, and that the creation of a meta variable boils down to the creation of a pointer that points to the current context. Inference of a meta type will still traverse the context, but this is already the case.

I think that this transformation should also allow to drop the `fresh` and `fresh_box` functions, so that we only work with the `make` functions. Consequently, environments `Env` should be used less as well, in favour of contexts `Ctxt`.