@fblanqui 2021-06-07
@gabrielhdt I see that you fixed the problem with https://github.com/fblanqui/lib but, now, there are a lot of new incorrect warnings like `[/home/blanqui/src/lambdapi/mylib/Eq.lp:26:15-17] Variable xy could be replaced by '_'.`