@gabrielhdt 2021-06-07
> @gabrielhdt I see that you fixed the problem with https://github.com/fblanqui/lib but, now, there are a lot of new incorrect warnings like `[/home/blanqui/src/lambdapi/mylib/Eq.lp:26:15-17] Variable xy could be replaced by '_'.`

Did I? Just by adding a `reflexivity` tactic you mean?

Warnings appear because the variables introduced by `assume` are not bound in the body. Before this PR, the variables were bound in the environment of the meta variables at scoping time. Now that we use placeholders, an `assume a` is scoped to a `refine \a, ?` where `?` is a placeholder, so `a` is indeed unbound. The old behaviour would do `refine \a, ?[a]`, so `a` would be bound.