@gabrielhdt 2021-06-07
> Could you please give more details about that? Where the context is traversed? Le 06/06/2021 à 14:30, Gabriel Hondet a écrit :
> […](#)
> Regarding performance problem, I think they are due to the fact that during typechecking, we traverse the context more often than before (twice as many times), since it is traversed to transform placeholders into meta variables and then to instantiate metavariables.

Hmm it's not as obvious as I thought, but for instance when a placeholder for a value is encountered, it is traversed twice to generate two new meta variables (for the type and the meta itself).