@fblanqui 2021-06-07
Le 07/06/2021 à 08:35, Gabriel Hondet a écrit :
>
>     @gabrielhdt <https://github.com/gabrielhdt> I see that you fixed
>     the problem with https://github.com/fblanqui/lib
>     <https://github.com/fblanqui/lib> but, now, there are a lot of new
>     incorrect warnings like
>     |[/home/blanqui/src/lambdapi/mylib/Eq.lp:26:15-17] Variable xy
>     could be replaced by '_'.|
>
> Did I? Just by adding a |reflexivity| tactic you mean?
>
No, I didn't change the files (and we a priori shouldn't anyway).


> Warnings appear because the variables introduced by |assume| are not 
> bound in the body. Before this PR, the variables were bound in the 
> environment of the meta variables at scoping time. Now that we use 
> placeholders, an |assume a| is scoped to a |refine \a, ?| where |?| is 
> a placeholder, so |a| is indeed unbound. The old behaviour would do 
> |refine \a, ?[a]|, so |a| would be bound.
>
How do you propose to solve this problem?