@gabrielhdt 2021-11-16
> Hi @gabrielhdt . Do you know why the CI continues to check for ocaml 4.07.1 ?

No, I don't.