@QGarchery 2021-11-24
I have some trouble testing this PR. I believe this is due to commit 72d6e858877d9bdcd4d9966c4fcf0303abdc0fd0 where some additional reduction is done. The file `assume.txt` from #693 shows what the issue is.