@gabrielhdt 2021-11-24
> I have some trouble testing this PR. I believe this is due to commit [72d6e85](https://github.com/Deducteam/lambdapi/commit/72d6e858877d9bdcd4d9966c4fcf0303abdc0fd0) where some additional reduction is done. The file `assume.txt` from #693 shows what the issue is.

Hi @QGarchery. After some bisection, it's the (not so) new type checking algorithm which causes your example to fail (commit 9ae0100d8a6346300f56988356fde83a54838685). I'm not surprised that this algorithm does a bit more reduction. I'll have a look.