@fblanqui 2021-11-30
Hi @gabrielhdt. I do not understand what is the problem here. Did you solve it?