@gabrielhdt 2021-11-30
> Hi @gabrielhdt. I do not understand what is the problem here. Did you solve it?

Not yet, but I think the problem is that this typechecker may perform reductions that the previous typechecker wouldn't perform. Some tactics seem to be a bit fragile, in the sense that they depend syntactically on the goal (and not modulo). It's a guess though. My idea is to try to not record reductions that are performed.